1st version

+++ sql/crud.py
@@ -1,13 +1,13 @@
-from sqlalchemy.orm import session
+from sqlalchemy.orm import Session
 from sql.models import Department, Employee
 
 
-def get_all_deptts(db: session):
+def get_all_deptts(db:Session):
     deptt = db.query(Department).all()
     return deptt
 
 
-def add_deptt(data: dict, db: session):
+def add_deptt(data: dict, db: Session):
     deptt = Department(**data)
     db.add(deptt)
     db.commit()
@@ -16,8 +16,8 @@ def add_deptt(data: dict, db: session):
     return deptt
 
 
-def update_deptt(data: dict, deptt_id: int, db: session):
-    deptt = db.query(Department).filter(Department.id = deptt_id).first()
+def update_deptt(data: dict, deptt_id: int, db: Session):
+    deptt = db.query(Department).filter(Department.id==deptt_id).first()
 
     for key, value in data.items():
         setattr(deptt, key, value)
@@ -27,19 +27,21 @@ def update_deptt(data: dict, deptt_id: int, db: session):
     return deptt
 
 
-def delete_deptt(deptt_id: int, db: session):
-    deptt = db.query(Department).filter(Department.id = deptt_id).first()
+def delete_deptt(deptt_id: int, db: Session):
+    deptt = db.query(Department).filter(Department.id==deptt_id).first()
+    if deptt is None:
+        return False
     db.delete(deptt)
     db.commit()
     return True
 
 
-def get_all_emps(db: session):
+def get_all_emps(db: Session):
     emp = db.query(Employee).all()
     return emp
 
 
-def add_emp(data: dict, db: session):
+def add_emp(data: dict, db: Session):
     emp = Employee(**data)
     db.add(emp)
     db.commit()
@@ -47,8 +49,8 @@ def add_emp(data: dict, db: session):
     return emp
 
 
-def update_emp(data: dict, deptt_id: int, db: session):
-    emp = = db.query(Employee).filter(Employee.id = emp_id).first()
+def update_emp(data: dict, emp_id: int, db: Session):
+    emp = db.query(Employee).filter(Employee.id==emp_id).first()
 
     for key, value in data.items():
         setattr(emp, key, value)
@@ -58,8 +60,10 @@ def update_emp(data: dict, deptt_id: int, db: session):
     return emp
 
 
-def delete_emp(emp_id: int, db: session):
-    emp = db.query(Employee).filter(Employee.id = emp_id).first()
+def delete_emp(emp_id: int, db: Session):
+    emp = db.query(Employee).filter(Employee.id==emp_id).first()
+    if emp is None:
+        return False
     db.delete(emp)
     db.commit()
     return True
